refactor command composition
