+++ app/render.py
@@ -1,5 +1,3 @@
-import os
-
 from emergency_alerts_utils.formatters import autolink_urls, formatted_list
 from jinja2 import (
     ChoiceLoader,
@@ -114,18 +112,3 @@ def get_rendered_pages(alerts):
         rendered["alerts/" + target] = template.render()
 
     return rendered
-
-
-def get_assets(folder):
-    assets = {}
-
-    for root, _, files in os.walk(folder):
-        s3path = root[root.find("alerts/"):]
-
-        # ignore hidden files and folders
-        files = [f for f in files if not f[0] == '.']
-
-        for file in files:
-            assets[s3path] = file
-
-    return assets
